Don't hand-craft avatar URLs, just use wAvatar's getAvatarURL() method for building the <img> tag
The generated <img> element now has:
* a non-empty alt attribute ("avatar"
* class="mw-socialprofile-avatar"
* Furthermore whether an avatar is or isn't output also depends on $wgUserProfileDisplay['avatar'], as some crazy sites out there might not want avatar and thus have $wgUserProfileDisplay['avatar'] = false; set

This is needed for the eventual FileBackend support, to ensure that if SocialProfile is using a non-default file backend for storing avatars (and/or other images), PollNY will use the correct functions for building the avatar URL instead of blindly assuming NFS when the wiki might be using Swift or somesuch instead.

Also gets rid of some inline CSS (ping T147257).

Change-Id: Iff56f745baf2c355e5da9c74479068ce1a24e428