Move formatting from PHP code to the i18n file(s) to fix an obvious PLURAL bug with the poll-vote-for i18n msg + adjust documentation
PLURAL wasn't working because $1 was not an integer but rather an integer wrapped in <b>...</b> and thus essentially interpreted as a string, or something.

Also the i18n msg in question is used on Poll: pages, below the "About the submitter" box, and *not* by the JS code; adjusted qqq accordingly.

Change-Id: I0d1137258d82a1872fd46f43287c28fd34f8f808