Correctly hide the "Skip" button on Poll: pages when JS is disabled
Apparently I didn't implement a no-JS skip handler back in the day as it's slightly complicated, but the CSS was there.
Thus, obviously, the CSS also didn't match the selectors which are currently used.
For the time being, just hide the "Skip" link for no-JS users; it's not like it's even the only feature in PollNY that requires JS.

Follow-up to bbe723ea2cc38fee0de00a50d032e7d2c68bc2d1.

Change-Id: I57ba957118444fd2fc4e2961e067ed912163c7d4